pyproject.toml
[project]
name = "forefore"
version = "0.1.0"
description = ""
authors = [
    {name = "Saemi",email = "cchocolatec@naver.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "django-split-settings (>=1.3.2,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)"
]

[tool.isort]
multi_line_output = 5
line_length = 119

[tool.yapf]
based_on_style = "google"
align_closing_bracket_with_visual_indent = true
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true

# mypy가 migrations 하위는 검사하지 않도록 설정
[tool.mypy]
# exclude = "core/.*/migrations/.*"
exclude = "core/products/migrations/0001_initial.py"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
